+++ src/main/java/app/boutique/metier/IProduitMetier.java
@@ -8,6 +8,6 @@ import app.boutique.dto.Produit;
 
 public interface IProduitMetier {
 	
-	public Produit getProduitById(long idProduit);
+	public Produit getProduitById(int idProduit);
 	public List<Produit> lisProduits();
 }
